+++ docs/conf.py
@@ -22,8 +22,8 @@ copyright = '2019, RTE France'
 author = 'Benjamin DONNOT'
 
 # The full version, including alpha/beta/rc tags
-release = "0.9.2.post2"
-version = '0.9'
+release = "0.10.0"
+version = '0.10'
 
 # -- General configuration ---------------------------------------------------